+++ modules/search-filters/search-filters.php
@@ -1,0 +1,7 @@
+<?php function telabotanica_module_search_filters($data) {
+  echo '<div class="search-filters">';
+	  echo '<h2 class="search-filters-title">' . __('Filtrer les résultats', 'telabotanica') . '</h2>';
+	  echo '<ul class="search-filters-items">';
+	  echo '</ul>';
+  echo '</div>';
+}