+++ modules/search-results/inc/hit-vegetation.php
@@ -1,4 +1,4 @@
-<?php function telabotanica_module_search_results_hit_syntaxons($hit) {
+<?php function telabotanica_module_search_results_hit_vegetation($hit) {
 	printf(
 		'<a class="search-results-hit-link" href="%s" target="_blank" title="%s">',
 		$hit['url'],